Experiment with how strict we need to be.
